refactor: player
